feat(product): 增加产品查询字段和移动历史记录及盘点类型文件
